fixup Index.name
