+++ doc/source/whatsnew/v1.0.0.rst
@@ -204,6 +204,7 @@ Deprecations
 - ``Index.set_value`` has been deprecated. For a given index ``idx``, array ``arr``,
   value in ``idx`` of ``idx_val`` and a new value of ``val``, ``idx.set_value(arr, idx_val, val)``
   is equivalent to ``arr[idx.get_loc(idx_val)] = val``, which should be used instead (:issue:`28621`).
+-
 
 .. _whatsnew_1000.prior_deprecations:
